pyproject.toml
[project]
name = "chat-assistant"
version = "0.1.0"
description = "A FastAPI Chat Assistant API with document RAG capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.111.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-groq>=0.1.0",
    "langchain-pinecone>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "pillow>=10.3.0",
    "pinecone-client>=4.1.0",
    "pyarrow>=23.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.26",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "fastembed>=0.3.0",
    "tornado>=6.5.5",
    "uvicorn>=0.30.0",
    "watchfiles>=1.1.1",
]
